Builder fusion setup

+++ api/db/migrate/20230821175749_drop_project_memberships.rb
@@ -1,5 +1,8 @@
 class DropProjectMemberships < ActiveRecord::Migration[7.0]
   def change
-    drop_table :project_memberships
+    # Skip if table doesn't exist yet
+    if table_exists?(:project_memberships)
+      drop_table :project_memberships
+    end
   end
 end
